A whole lotta front end stuff ive been working on
I made it super slick and sexy, the HTML is super self-explanatory so attaching php to it is not difficult. Just some basic forms and inputs and buttons. HOWEVER, later ill be having either Anthony or I to fully integrate media queries as it works perfectly fine on LARGER screens but on smaller ones it begins to act wacky. That isnt right now the priority, as ill start working on the other pages. Those being a create post page, the dashboard, profile page and so on.