Add files via upload
Ok, this file is for the whole database so you will have to delete your current database (maindb) and import this file. The file upload will only work if you have a folder called uploads inside the projectUofT. Each user has a subfolder inside the uploads folder. This folder is created on signup automatically with the username, so you will have to create these subfolders for pre-existing users.